Use the CodeBase for the assembly to determine where resources are.
Original code used Assembly.Location, which is where the assembly currently is. However, when using the nUnit ShadowCache option, it copies the binaries to a temporary folder %TEMP%\nunit20\ShadowCopyCache but doesn't copy the deployment items. The solution as per the blog https://blog.jetbrains.com/dotnet/2015/12/28/shadow-copying-in-dotcover-if-your-nunit-tests-fail-during-continuous-testing/ is to use the property CodeBase. That knows where the assembly was compiled to originally where the source objects are kept.

Tested on Windows 10 with nUnit desktop app with shadow copy on and off. The DeploymentItem test cases pass.