Rework DeploymentItemAttribute to handle parallel copying by retrying.
On slow disks we have a much higher chance of running into an exception of trying to copy a file that is already in the process of being copied. That occurs because after what's on disk is not what we would expect, two methods then start copying. Typical errors are listed in code. We don't want a global lock, so we catch the exceptions that occur, sleep 250ms and try again. We might try again if the file exists. Hopefully the copy is finished within 1 second and all is good. If it isn't, we ignore it and continue. Why do we ignore it? Either the user had an error and the results are undefined anyway (see the summary remarks), or it's already copied.

Further change includes trimming the output directory. If it contains spaces at the end or beginning, we remove them now.

Tested by using the console version of nUnit which makes exceptions really obvious (it still doesn't print to the console though). Tested on a really slow disk that showed the problems, and on a fast SSD that never showed any problems.

Issue: MGUSIN-7984