Refactor to allow deployment of files using a method as well as an attribute.
The method allows for more flexibility. One use case is to deploy to a directory that is calculated as part of the test case.

This method would only be executed when the test case is run. So when test cases switch over to use this method, deployment only occurs when the test case is run, not during reflection.

Reviewed-by: Adrian Constantin Vaideanu <avaideanu@luxoft.com>