Change the way that Deploy.Item() works to be more efficient.
Deploy.Item() now compares folders when copying and only copies where there's a difference, also deleting files in the directory, if it shouldn't be there.

We discovered a bug with the deployment of some test cases that used an old path "Data" instead of the new "TestDeployment", which was found by the changes.

Provide a base class NUnitExtensions that one can derive from for automatic deployment.

Fix handling of the Assembly CodeBase on Linux. Properly convert the file:// URI to a locl path, before getting the directory name. Getting the directory name beforehand would strip and modify the URI that it was no longer valid, causing it to fail on Linux when using Mono.

Issue: DOTNET-13