Testing internal classes support: test generic classes, provide base accessor class.
Added support for testing internal generic classes by using a PrivateObject to have a similar API to MSTest framework. Added tests for accessing generic classes.

Added base accessor class which wraps an instance of PrivateObject used for testing internal classes or accessing private members. The purpose of the accessor is to hide the usage of the PrivateObject and to provide functionalities for defining similar signatures to the class which is tested. The accessor can be used for testing both generic and non-generic classes. Added test scenarios for accessing public/private members, raising events for generic and non-generic classes through the base accessor.

Added overloaded constructor for PrivateObject in order to be able to call overloaded methods. This will allow to set the value of the bindingFlags only once in the constructor of the accessor, and still be able to call public and private methods. In this way accessing both public and private properties will not need to modify the value of the bindingFlags in getter/setter. Added test cases for calling public/protected overloaded methods. Removed duplicate value for the BindingFlags in PrivateType class.

Calling an inexistent method or property on a PrivateObject throws MissingMethodException as it does in MSTest.

Ran all unit tests.

Issue: HELIOS-210