Expand implementation of PrivateObject and PrivateType to be compatible with MS implementation
Not all methods have been transferred, but the now PrivateType and PrivateObject contain all the methods for invoking, setting and getting fields or properties.

Test cases have been created that can test the MS implementation and the NUnitExtensions implementation with the same test code for direct comparison to enable the ability to ensure that the NUnitExtensions implementation behaves the same as the MS implementation. Only the test cases depend on the reference to Microsoft.VisualStudio.QualityTools.UnitTestFramework, the NUnitExtensions assembly doesn't, allowing it to be deployed.

The NUnitExtensions implementation has been implemented by decompiling v10.1.0.0 of Microsoft.VisualStudio.QualityTools.UnitTestFramework, then integrating into this library. There are a few minor differences, such as how exceptions are raised (stack frames will differ between the two implementations) and the exceptions raised when calling PrivateObject and PrivateType will slightly differ (the exception types should be the same however). The advantages of using this implementation is for an easier deployment.

NOTE: This code contains portions copyright Microsoft through decompilation with dotPeek, and should not be deployed outside of a private environment.

All methods in PrivateObject/PrivateType with culture in the signature are removed, making slightly different interfaces to the MS implementation. The culture doesn't appear to have any effect on testing at all. Indeed, test cases written against the MS version and our version to set the culture to English and German showed no visible effects and reading the culture was still the same as the thread's default culture.

The culture is only used when calling Invoke, InvokeMember, MakeGenericMethod. Microsoft Reference Sources, show that it is only given to the Binder (which is always null in PrivateObject/PrivateTest). As the Binder is null, the default binder is used and doesn't have any culture specific type coercion. Thus it appears, MS implemented the PrivateObject/PrivateType with a lot of dead code.

* Implemented a new class GenericPrivateObject

Refactored all generic tests into the GenericPrivateObject. For this to work, we use generic types in the test case, and then had to create classes that derive from the same interface and wrap the Microsoft object and our object so we can compile without duplicating test code.

MS PrivateObject has no way to create private objects. You have to create the type yourself in this case. To reduce the effort required, a new type GenericPrivateObject has a simple constructor to create the PrivateObject we want. That way, we can instantiate this object whenever we want in place of the PrivateObject, and it will still work as it did before.

Added GenericPrivateType, which is functionality missing from original MS PrivateType, as a replacement for some of the functionality being lost in this commit.

Issue: HELIOS-224