PrivateObject and PrivateType now always raise TargetInvocationException
The implementation of PrivateObject and PrivateType will sometimes raise the exception of a class under test, or sometimes raise the TargetInvocationException depending if parameterTypes is non-null or null. This is how VS2015 implemented it. Now, always raise the TargetInvocationException and make behaviour consistent.

The interfaces to AccessorBase have not changed, they will not raise TargetInvocationException.

Added more unit test cases for better coverage in AccessorBase.

Issue: HELIOS-255