NUnitExtensions: Speed up File and Directory Delete operations
Windows may have other processes that have a file open, which prevent it from being deleted (or the directory being deleted), which requires a retry mechanism. This is implemented using the FileSystemWatcher class by Microsoft.

On Mono platforms, this was being used also, but the FileSystemWatcher wasn't triggering the ManualResetEvent when the file/directory was being deleted, thus resulting in each operation taking 250ms. As the usage of the FileSystemWatcher is only a workaround for Windows systems, we remove its implementation completely for Unix based systems, which can delete a file/directory even if there is a file handle open.

Issue: HELIOS-433