NUnitExtensions: Upgrade to NUnit 2.6.7 and remove DeploymentItemAttribute
When moving to NUnit 2.6.7, the [TestFixtureSetUp] attribute should be replaced with [OneTimeSetUp]. Unfortunately, it doesn't work and if we move to NUnit 3.x (which is the reason to moving to NUnit 2.6.7 earlier) there is no solution. Given that the [OneTimeSetUp] doesn't actually get executed, remove the entire concept of the base class.

By removing the base class, test cases will need to explicitly deploy files with Deploy.Item(). This is arguably better, as it is already observed some test cases use the [DeploymentItem] attribute, but forget to derive from NUnitExtensions (which is an error, and would result in no deployment).

This project no longer has a dependency to NUnit since the removal of the base class NUnitExtensions and so has been removed.

Issue: HELIOS-548