NUnitExtensions: Allow the possibility to create a crash dump for deep debugging
Automated tests need a way to dump the state, and then continue without user interaction. In C, usually one uses an "Assert" which creates a SIGABORT and then can be used to later analyze a crash dump. The user can force a minidump which can be loaded with Visual Studio later and software continues its execution.

Issue: HELIOS-785