NUnitExtensions: Provide TestDirectory and WorkDirectory from NUnit.Framework
Without adding a reference to a version of NUnit, use reflection to get the current test assembly and obtain the TestDirectory and WorkDirectory.

The behaviour of the Deploy class has changed as follows:
- Deploy.Item source is now relative to TestDirectory, destination is relative to WorkDirectory.
- Deploy.CreateDirectory is relative to WorkDirectory.
- Deploy.DeleteFile is relative to WorkDirectory.
- Deploy.DeleteDirectory is relative to WorkDirectory.

The path to WorkDirectory can be modified in the App.Config file, which is especially useful when testing correctness within the Visual Studio IDE (it ensures that TestDirectory and WorkDirectory are different, which is often the case when using the console runner, but not when using the IDE).

  <configSections>
    <section name="NUnitExtensions" type="NUnit.Framework.AppConfig.NUnitExtensionsSection, RJCP.NUnitExtensions"/>
  </configSections>

  <NUnitExtensions>
   <deploy workDir="work" force="false"/>
  </NUnitExtensions>

For backwards compatibility with NUnit 2.x test cases where the runner is started with the option /work specified, use the following configuration to deploy to the current directory, instead of the work directory. This will not work with NUnit 3, as the current directory using NUnit 3 in the Visual Studio IDE is not writable.

  <configSections>
    <section name="NUnitExtensions" type="NUnit.Framework.AppConfig.NUnitExtensionsSection, RJCP.NUnitExtensions"/>
  </configSections>

  <NUnitExtensions>
   <deploy useCwd="true"/>
  </NUnitExtensions>

Issue: HELIOS-865