NUnitExtensions: Look for NUnit when not available in the stack trace
When programming with Async and threads, a test case might require the TestContextAccessor, but the stack trace cannot be used to determine the type. In this scenario, iterate through all the open assemblies and look for the type.

This of course will only work if there is only one NUnit assembly containing the TestAttribute.

Issue: HELIOS-1301