RJCP.CodeQuality: Rename from NUnitExtensions
The assembly has grown to be much more than simple extensions for helping with NUnit. It contains methods that do more.

Rename the project to be RJCP.CodeQuality, with NUnitExtensions being a part of the namespace. Items that have nothing to do with NUnit are out of the namespace to make it clear it's generic.

Issue: HELIOS-1337