ScratchPad: Allow the scratchpad to be used twice in the same test case
A user may want to initialize the scratchpad twice in the same test case. This might be necessary, if for example, instantiating the scratch pad is hidden in another class.

Note, with the default options, the directory will be erased each time.

The function to convert the test case name to a directory name was updated to remove ending underscores '_'.

Issue: DOTNET-264