CodeQuality: Upgrade to build for .NET Core
The project file was updated to use the .NET SDK, which allows targetting multiple frameworks.

The code now supports .NET Core using .NET Standard 2.1. The logic in loading a configuration file from the `app.config` file needed a new implementation due to a lack of full support from .NET (ConfigurationManager.GetSection() doesn't work well in .NET Core for loading the default configuration for test assemblies).

The CodeQualityTest.NUnit2.csproj file is the same as the main test assembly, but the dotnet toolchain requires that the .csproj file be in a different folder (because when it compiles two .csproj files in the same folder, it copies it still to the same path and thus it might confuse itself as the generated metadata differs between the two projects, and may result in inconsistent builds).

This means the resource files must be duplicated, as the project looks for them in the original location.

Issue: DOTNET-343, DOTNET-344