NUnitExtensions: Add a NUnitLogger
Add an NUnitLogger, based on the example given in [1], and documentation from Microsoft in [2]. When logging, it might be tempting to use `.AddConsole()`, but NUnit can't log the console reliably, and results in missing output, or output in the wrong test case. Using NUnit.TestContext.Write() resolves this problem and allows your test cases to log to NUnit.

[1] https://github.com/nunit/nunit/issues/3919#issuecomment-904194817
[2] https://docs.microsoft.com/en-us/dotnet/core/extensions/custom-logging-provider

The code itself doesn't have a direct dependency to NUnit.Framework, but pulls it in via reflection. This allows the user to have the library automaticall adapt to NUnit 2.x or NUnit 3.x (they're similar, but not compatible and otherwise both can't be used at the same time).

Issue: DOTNET-382