Streams: Replace ArgumentOutOfRangeException with ThrowHelper
Some fixes were made to `SimpleStream.Seek` to throw `ArgumentOutOfRangeException` instead of `IOException` or `ArgumentException`, as well as now properly checking the input of `offset`.

Fixes for `SparseStream.Seek` to raise the correct exception if we seek beyond the maximum value of the position. Setting the position out of range now throws `ArgumentOutOfRangeException` instead of `ArgumentException`.

Issue: DOTNET-989